+++ tableView.go
@@ -606,7 +606,8 @@ func (table *tableViewData) propertyChanged(tag string) {
 
 			switch GetTableSelectionMode(table) {
 			case CellSelection:
-				session.updateProperty(htmlID, "tabindex", "0")
+				tabIndex, _ := intProperty(table, TabIndex, session, 0)
+				session.updateProperty(htmlID, "tabindex", tabIndex)
 				session.updateProperty(htmlID, "onfocus", "tableViewFocusEvent(this, event)")
 				session.updateProperty(htmlID, "onblur", "tableViewBlurEvent(this, event)")
 				session.updateProperty(htmlID, "data-selection", "cell")
@@ -621,7 +622,8 @@ func (table *tableViewData) propertyChanged(tag string) {
 				session.updateProperty(htmlID, "onkeydown", "tableViewCellKeyDownEvent(this, event)")
 
 			case RowSelection:
-				session.updateProperty(htmlID, "tabindex", "0")
+				tabIndex, _ := intProperty(table, TabIndex, session, 0)
+				session.updateProperty(htmlID, "tabindex", tabIndex)
 				session.updateProperty(htmlID, "onfocus", "tableViewFocusEvent(this, event)")
 				session.updateProperty(htmlID, "onblur", "tableViewBlurEvent(this, event)")
 				session.updateProperty(htmlID, "data-selection", "row")
@@ -636,7 +638,11 @@ func (table *tableViewData) propertyChanged(tag string) {
 				session.updateProperty(htmlID, "onkeydown", "tableViewRowKeyDownEvent(this, event)")
 
 			default: // NoneSelection
-				for _, prop := range []string{"tabindex", "data-current", "onfocus", "onblur", "onkeydown", "data-selection"} {
+				if tabIndex, ok := intProperty(table, TabIndex, session, -1); !ok || tabIndex < 0 {
+					session.removeProperty(htmlID, "tabindex")
+				}
+
+				for _, prop := range []string{"data-current", "onfocus", "onblur", "onkeydown", "data-selection"} {
 					session.removeProperty(htmlID, prop)
 				}
 			}
@@ -831,7 +837,7 @@ func (table *tableViewData) htmlSubviews(self View, buffer *strings.Builder) {
 	var view tableCellView
 	view.init(session)
 
-	ignorCells := []struct{ row, column int }{}
+	ignoreCells := []struct{ row, column int }{}
 	selectionMode := GetTableSelectionMode(table)
 
 	var allowCellSelection TableAllowCellSelection = nil
@@ -908,7 +914,7 @@ func (table *tableViewData) htmlSubviews(self View, buffer *strings.Builder) {
 
 			for column := 0; column < columnCount; column++ {
 				ignore := false
-				for _, cell := range ignorCells {
+				for _, cell := range ignoreCells {
 					if cell.row == row && cell.column == column {
 						ignore = true
 						break
@@ -994,7 +1000,7 @@ func (table *tableViewData) htmlSubviews(self View, buffer *strings.Builder) {
 						buffer.WriteString(strconv.Itoa(columnSpan))
 						buffer.WriteRune('"')
 						for c := column + 1; c < column+columnSpan; c++ {
-							ignorCells = append(ignorCells, struct {
+							ignoreCells = append(ignoreCells, struct {
 								row    int
 								column int
 							}{row: row, column: c})
@@ -1010,7 +1016,7 @@ func (table *tableViewData) htmlSubviews(self View, buffer *strings.Builder) {
 						}
 						for r := row + 1; r < row+rowSpan; r++ {
 							for c := column; c < column+columnSpan; c++ {
-								ignorCells = append(ignorCells, struct {
+								ignoreCells = append(ignoreCells, struct {
 									row    int
 									column int
 								}{row: r, column: c})